Move SceneValidators to a separate folder.
