+++ SceneryWebApi/Scenery.Controllers/Validators/SceneValidators/IntersectedSceneValidator.cs
@@ -1,24 +1,24 @@
-﻿// <copyright file="UnitedSceneValidator.cs" company="dsnouck">
+﻿// <copyright file="IntersectedSceneValidator.cs" company="dsnouck">
 // Copyright (c) dsnouck. All rights reserved.
 // Licensed under the MIT License. See LICENSE file in the project root for full license information.
 // </copyright>
 
-namespace Scenery.Controllers.Validators
+namespace Scenery.Controllers.Validators.SceneValidators
 {
     using FluentValidation;
     using Scenery.Components.Interfaces;
     using Scenery.Models.Scenes;
 
     /// <summary>
-    /// Represents a validator for <see cref="UnitedScene"/>.
+    /// Represents a validator for <see cref="IntersectedScene"/>.
     /// </summary>
-    public class UnitedSceneValidator : AbstractValidator<UnitedScene>
+    public class IntersectedSceneValidator : AbstractValidator<IntersectedScene>
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="UnitedSceneValidator"/> class.
+        /// Initializes a new instance of the <see cref="IntersectedSceneValidator"/> class.
         /// </summary>
         /// <param name="vector3Component">An <see cref="IVector3Component"/>.</param>
-        public UnitedSceneValidator(IVector3Component vector3Component)
+        public IntersectedSceneValidator(IVector3Component vector3Component)
         {
             this.RuleFor(scene => scene.Scenes)
                 .NotNull();
