--- SceneryWebApi/Scenery.ControllersTests/Validators/IntersectedSceneValidatorTests.cs
@@ -1,85 +1,0 @@
-﻿// <copyright file="IntersectedSceneValidatorTests.cs" company="dsnouck">
-// Copyright (c) dsnouck. All rights reserved.
-// Licensed under the MIT License. See LICENSE file in the project root for full license information.
-// </copyright>
-
-namespace Scenery.ControllersTests.Validators
-{
-    using System.Collections.Generic;
-    using FluentValidation.TestHelper;
-    using Moq;
-    using Scenery.Components.Interfaces;
-    using Scenery.Controllers.Validators;
-    using Scenery.Models;
-    using Scenery.Models.Scenes;
-    using Xunit;
-
-    /// <summary>
-    /// Provides tests for <see cref="IntersectedSceneValidator"/>.
-    /// </summary>
-    public class IntersectedSceneValidatorTests
-    {
-        private readonly SceneContainerValidator systemUnderTest;
-        private readonly Mock<IVector3Component> vector3ComponentTestDouble;
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="IntersectedSceneValidatorTests"/> class.
-        /// </summary>
-        public IntersectedSceneValidatorTests()
-        {
-            this.vector3ComponentTestDouble = new Mock<IVector3Component>();
-            this.vector3ComponentTestDouble
-                .Setup(vector3Component => vector3Component.GetLength(It.IsAny<Vector3>()))
-                .Returns(1D);
-
-            this.systemUnderTest = new SceneContainerValidator(this.vector3ComponentTestDouble.Object);
-        }
-
-        /// <summary>
-        /// Tests <see cref="IntersectedSceneValidator"/>.
-        /// </summary>
-        [Fact]
-        public void GivenScenesIsNullWhenValidateIsCalledThenItFails()
-        {
-            // Arrange.
-            var sceneContainer = new SceneContainer
-            {
-                Scene = new IntersectedScene
-                {
-                    Scenes = null,
-                },
-            };
-
-            // Act.
-            var result = this.systemUnderTest.TestValidate(sceneContainer);
-
-            // Assert.
-            result.ShouldHaveValidationErrorFor(sceneContainer => (sceneContainer.Scene as IntersectedScene).Scenes);
-        }
-
-        /// <summary>
-        /// Tests <see cref="IntersectedSceneValidator"/>.
-        /// </summary>
-        [Fact]
-        public void GivenScenesContainsNullWhenValidateIsCalledThenItFails()
-        {
-            // Arrange.
-            var sceneContainer = new SceneContainer
-            {
-                Scene = new IntersectedScene
-                {
-                    Scenes = new List<Scene>
-                    {
-                        null,
-                    },
-                },
-            };
-
-            // Act.
-            var result = this.systemUnderTest.TestValidate(sceneContainer);
-
-            // Assert.
-            result.ShouldHaveValidationErrorFor(sceneContainer => (sceneContainer.Scene as IntersectedScene).Scenes);
-        }
-    }
-}
