+++ SceneryWebApi/Scenery.ControllersTests/Validators/SceneValidators/InvertedSceneValidatorTests.cs
@@ -1,30 +1,31 @@
-﻿// <copyright file="SceneValidatorTests.cs" company="dsnouck">
+﻿// <copyright file="InvertedSceneValidatorTests.cs" company="dsnouck">
 // Copyright (c) dsnouck. All rights reserved.
 // Licensed under the MIT License. See LICENSE file in the project root for full license information.
 // </copyright>
 
-namespace Scenery.ControllersTests.Validators
+namespace Scenery.ControllersTests.Validators.SceneValidators
 {
     using FluentValidation.TestHelper;
     using Moq;
     using Scenery.Components.Interfaces;
     using Scenery.Controllers.Validators;
+    using Scenery.Controllers.Validators.SceneValidators;
     using Scenery.Models;
     using Scenery.Models.Scenes;
     using Xunit;
 
     /// <summary>
-    /// Provides tests for <see cref="SceneValidator"/>.
+    /// Provides tests for <see cref="InvertedSceneValidator"/>.
     /// </summary>
-    public class SceneValidatorTests
+    public class InvertedSceneValidatorTests
     {
         private readonly SceneContainerValidator systemUnderTest;
         private readonly Mock<IVector3Component> vector3ComponentTestDouble;
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="SceneValidatorTests"/> class.
+        /// Initializes a new instance of the <see cref="InvertedSceneValidatorTests"/> class.
         /// </summary>
-        public SceneValidatorTests()
+        public InvertedSceneValidatorTests()
         {
             this.vector3ComponentTestDouble = new Mock<IVector3Component>();
             this.vector3ComponentTestDouble
@@ -35,22 +36,25 @@ namespace Scenery.ControllersTests.Validators
         }
 
         /// <summary>
-        /// Tests <see cref="SceneValidator"/>.
+        /// Tests <see cref="InvertedSceneValidator"/>.
         /// </summary>
         [Fact]
-        public void GivenSceneIsOfTypeSceneWhenValidateIsCalledThenItFails()
+        public void GivenOriginalSceneIsNullWhenValidateIsCalledThenItFails()
         {
             // Arrange.
             var sceneContainer = new SceneContainer
             {
-                Scene = new Scene(),
+                Scene = new InvertedScene
+                {
+                    OriginalScene = null,
+                },
             };
 
             // Act.
             var result = this.systemUnderTest.TestValidate(sceneContainer);
 
             // Assert.
-            result.ShouldHaveValidationErrorFor(sceneContainer => sceneContainer.Scene);
+            result.ShouldHaveValidationErrorFor(sceneContainer => (sceneContainer.Scene as InvertedScene).OriginalScene);
         }
     }
 }
