+++ SceneryWebApi/Scenery.ControllersTests/Validators/SceneValidators/PlaneSceneValidatorTests.cs
@@ -1,30 +1,31 @@
-﻿// <copyright file="ScaledSceneValidatorTests.cs" company="dsnouck">
+﻿// <copyright file="PlaneSceneValidatorTests.cs" company="dsnouck">
 // Copyright (c) dsnouck. All rights reserved.
 // Licensed under the MIT License. See LICENSE file in the project root for full license information.
 // </copyright>
 
-namespace Scenery.ControllersTests.Validators
+namespace Scenery.ControllersTests.Validators.SceneValidators
 {
     using FluentValidation.TestHelper;
     using Moq;
     using Scenery.Components.Interfaces;
     using Scenery.Controllers.Validators;
+    using Scenery.Controllers.Validators.SceneValidators;
     using Scenery.Models;
     using Scenery.Models.Scenes;
     using Xunit;
 
     /// <summary>
-    /// Provides tests for <see cref="ScaledSceneValidator"/>.
+    /// Provides tests for <see cref="PlaneSceneValidator"/>.
     /// </summary>
-    public class ScaledSceneValidatorTests
+    public class PlaneSceneValidatorTests
     {
         private readonly SceneContainerValidator systemUnderTest;
         private readonly Mock<IVector3Component> vector3ComponentTestDouble;
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="ScaledSceneValidatorTests"/> class.
+        /// Initializes a new instance of the <see cref="PlaneSceneValidatorTests"/> class.
         /// </summary>
-        public ScaledSceneValidatorTests()
+        public PlaneSceneValidatorTests()
         {
             this.vector3ComponentTestDouble = new Mock<IVector3Component>();
             this.vector3ComponentTestDouble
@@ -35,17 +36,17 @@ namespace Scenery.ControllersTests.Validators
         }
 
         /// <summary>
-        /// Tests <see cref="ScaledSceneValidator"/>.
+        /// Tests <see cref="PlaneSceneValidator"/>.
         /// </summary>
         [Fact]
-        public void GivenOriginalSceneIsNullWhenValidateIsCalledThenItFails()
+        public void GivenNormalIsNullWhenValidateIsCalledThenItFails()
         {
             // Arrange.
             var sceneContainer = new SceneContainer
             {
-                Scene = new ScaledScene
+                Scene = new PlaneScene
                 {
-                    OriginalScene = null,
+                    Normal = null,
                 },
             };
 
@@ -53,7 +54,7 @@ namespace Scenery.ControllersTests.Validators
             var result = this.systemUnderTest.TestValidate(sceneContainer);
 
             // Assert.
-            result.ShouldHaveValidationErrorFor(sceneContainer => (sceneContainer.Scene as ScaledScene).OriginalScene);
+            result.ShouldHaveValidationErrorFor(sceneContainer => (sceneContainer.Scene as PlaneScene).Normal);
         }
     }
 }
