+++ SceneryWebApi/Scenery.ControllersTests/Validators/SceneValidators/ScaledSceneValidatorTests.cs
@@ -1,30 +1,31 @@
-﻿// <copyright file="InvertedSceneValidatorTests.cs" company="dsnouck">
+﻿// <copyright file="ScaledSceneValidatorTests.cs" company="dsnouck">
 // Copyright (c) dsnouck. All rights reserved.
 // Licensed under the MIT License. See LICENSE file in the project root for full license information.
 // </copyright>
 
-namespace Scenery.ControllersTests.Validators
+namespace Scenery.ControllersTests.Validators.SceneValidators
 {
     using FluentValidation.TestHelper;
     using Moq;
     using Scenery.Components.Interfaces;
     using Scenery.Controllers.Validators;
+    using Scenery.Controllers.Validators.SceneValidators;
     using Scenery.Models;
     using Scenery.Models.Scenes;
     using Xunit;
 
     /// <summary>
-    /// Provides tests for <see cref="InvertedSceneValidator"/>.
+    /// Provides tests for <see cref="ScaledSceneValidator"/>.
     /// </summary>
-    public class InvertedSceneValidatorTests
+    public class ScaledSceneValidatorTests
     {
         private readonly SceneContainerValidator systemUnderTest;
         private readonly Mock<IVector3Component> vector3ComponentTestDouble;
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="InvertedSceneValidatorTests"/> class.
+        /// Initializes a new instance of the <see cref="ScaledSceneValidatorTests"/> class.
         /// </summary>
-        public InvertedSceneValidatorTests()
+        public ScaledSceneValidatorTests()
         {
             this.vector3ComponentTestDouble = new Mock<IVector3Component>();
             this.vector3ComponentTestDouble
@@ -35,7 +36,7 @@ namespace Scenery.ControllersTests.Validators
         }
 
         /// <summary>
-        /// Tests <see cref="InvertedSceneValidator"/>.
+        /// Tests <see cref="ScaledSceneValidator"/>.
         /// </summary>
         [Fact]
         public void GivenOriginalSceneIsNullWhenValidateIsCalledThenItFails()
@@ -43,7 +44,7 @@ namespace Scenery.ControllersTests.Validators
             // Arrange.
             var sceneContainer = new SceneContainer
             {
-                Scene = new InvertedScene
+                Scene = new ScaledScene
                 {
                     OriginalScene = null,
                 },
@@ -53,7 +54,7 @@ namespace Scenery.ControllersTests.Validators
             var result = this.systemUnderTest.TestValidate(sceneContainer);
 
             // Assert.
-            result.ShouldHaveValidationErrorFor(sceneContainer => (sceneContainer.Scene as InvertedScene).OriginalScene);
+            result.ShouldHaveValidationErrorFor(sceneContainer => (sceneContainer.Scene as ScaledScene).OriginalScene);
         }
     }
 }
