+++ SceneryWebApi/Scenery.ControllersTests/Validators/SceneValidators/SceneValidatorTests.cs
@@ -1,30 +1,31 @@
-﻿// <copyright file="PlaneSceneValidatorTests.cs" company="dsnouck">
+﻿// <copyright file="SceneValidatorTests.cs" company="dsnouck">
 // Copyright (c) dsnouck. All rights reserved.
 // Licensed under the MIT License. See LICENSE file in the project root for full license information.
 // </copyright>
 
-namespace Scenery.ControllersTests.Validators
+namespace Scenery.ControllersTests.Validators.SceneValidators
 {
     using FluentValidation.TestHelper;
     using Moq;
     using Scenery.Components.Interfaces;
     using Scenery.Controllers.Validators;
+    using Scenery.Controllers.Validators.SceneValidators;
     using Scenery.Models;
     using Scenery.Models.Scenes;
     using Xunit;
 
     /// <summary>
-    /// Provides tests for <see cref="PlaneSceneValidator"/>.
+    /// Provides tests for <see cref="SceneValidator"/>.
     /// </summary>
-    public class PlaneSceneValidatorTests
+    public class SceneValidatorTests
     {
         private readonly SceneContainerValidator systemUnderTest;
         private readonly Mock<IVector3Component> vector3ComponentTestDouble;
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="PlaneSceneValidatorTests"/> class.
+        /// Initializes a new instance of the <see cref="SceneValidatorTests"/> class.
         /// </summary>
-        public PlaneSceneValidatorTests()
+        public SceneValidatorTests()
         {
             this.vector3ComponentTestDouble = new Mock<IVector3Component>();
             this.vector3ComponentTestDouble
@@ -35,25 +36,22 @@ namespace Scenery.ControllersTests.Validators
         }
 
         /// <summary>
-        /// Tests <see cref="PlaneSceneValidator"/>.
+        /// Tests <see cref="SceneValidator"/>.
         /// </summary>
         [Fact]
-        public void GivenNormalIsNullWhenValidateIsCalledThenItFails()
+        public void GivenSceneIsOfTypeSceneWhenValidateIsCalledThenItFails()
         {
             // Arrange.
             var sceneContainer = new SceneContainer
             {
-                Scene = new PlaneScene
-                {
-                    Normal = null,
-                },
+                Scene = new Scene(),
             };
 
             // Act.
             var result = this.systemUnderTest.TestValidate(sceneContainer);
 
             // Assert.
-            result.ShouldHaveValidationErrorFor(sceneContainer => (sceneContainer.Scene as PlaneScene).Normal);
+            result.ShouldHaveValidationErrorFor(sceneContainer => sceneContainer.Scene);
         }
     }
 }
